requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
wincertstore==0.2
